op/requirements.txt
Django==2.0.8
django-debug-toolbar==1.9.1
django-oidc-provider==0.6.2
